+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Add utility functions for dealing with common input ([#759])
 * Add alpha cutoff support to the PBR shader ([#756])
 * Basic renderer setup helper function ([#771])
+* Shape mesh generators ([#777])
 
 ### Changed
 * Refactored asset loading so `Processor`s can defer storage insertion ([#760])
@@ -46,6 +47,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#772]: https://github.com/amethyst/amethyst/pull/772
 [#773]: https://github.com/amethyst/amethyst/pull/773
 [#771]: https://github.com/amethyst/amethyst/pull/771
+[#777]: https://github.com/amethyst/amethyst/pull/777
 
 ## [0.7.0] - 2018-05
 ### Added